Pianta/requirements.txt
amqp==5.1.1
asgiref==3.6.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
crispy-bootstrap5==0.7
cron-descriptor==1.2.35
cryptography==40.0.1
defusedxml==0.7.1
dj-rest-auth==3.0.0
Django==4.1.7
django-allauth==0.53.1
django-celery-beat==2.5.0
django-crispy-forms==2.0
django-otp==1.1.6
django-random-id-model==0.1.1
django-resized==1.0.2
django-rest-framework==0.1.0
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
httpie==3.2.1
idna==3.4
kombu==5.2.4
markdown-it-py==2.2.0
mdurl==0.1.2
multidict==6.0.4
mysqlclient==2.1.1
oauthlib==3.2.2
Pillow==9.4.0
prompt-toolkit==3.0.38
pycparser==2.21
Pygments==2.14.0
PyJWT==2.6.0
PyMySQL==1.0.2
pyotp==2.8.0
PySocks==1.7.1
python-crontab==2.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rich==13.3.2
sendmail==2.0
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
